Bootcamps/Microsoft Certification Challenge #5 - AI 102/Analise-de-Documentos-Anti-fraude-com-AzureAI/requirements.txt
azure-ai-formrecognizer==3.3.3
python-dotenv==1.2.1
streamlit==1.54.0
pandas==2.2.3
